containers/eo/requirements.txt
azure-storage-blob
python-dotenv
